requirement.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
docker==5.0.3
fastapi==0.68.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
PyMySQL==1.0.2
#pywin32==227
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
websocket-client==1.2.1
